Fix exit status check in #run!
`#zero?` only works on integers.  `Open3`'s status is a
`Process::Status` object and surfaces a `NoMethodError`
when you try to call `#zero?` on it.

So assign `status.exitstatus` to the `status` variable during the
`quiet` run operation.

Wrote a few tests that will verify this continues working in the future